Remove 400 compiler warnings

+++ src/AABB.h
@@ -28,7 +28,7 @@ class AABB {
   private:
     // stores x,y,z co-ordinates
     struct XYZ {
-        GLdouble x = 0.0, y = 0.0, z = 0.0;
+        GLfloat x = 0.0, y = 0.0, z = 0.0;
     };
     // stores max and min values of co-ordinates
     struct BoundingBox {
@@ -57,22 +57,22 @@ class AABB {
     //----------------------------------------------------------------------------------
     //  Set Methods
     //----------------------------------------------------------------------------------
-    void SetMaxX(const int &tempIndex, const GLdouble &tempX) {
+    void SetMaxX(const int &tempIndex, const GLfloat &tempX) {
         m_BBox[tempIndex].max.x = tempX;
     }
-    void SetMinX(const int &tempIndex, const GLdouble &tempX) {
+    void SetMinX(const int &tempIndex, const GLfloat &tempX) {
         m_BBox[tempIndex].min.x = tempX;
     }
-    void SetMaxY(const int &tempIndex, const GLdouble &tempY) {
+    void SetMaxY(const int &tempIndex, const GLfloat &tempY) {
         m_BBox[tempIndex].max.y = tempY;
     }
-    void SetMinY(const int &tempIndex, const GLdouble &tempY) {
+    void SetMinY(const int &tempIndex, const GLfloat &tempY) {
         m_BBox[tempIndex].min.y = tempY;
     }
-    void SetMaxZ(const int &tempIndex, const GLdouble &tempZ) {
+    void SetMaxZ(const int &tempIndex, const GLfloat &tempZ) {
         m_BBox[tempIndex].max.z = tempZ;
     }
-    void SetMinZ(const int &tempIndex, const GLdouble &tempZ) {
+    void SetMinZ(const int &tempIndex, const GLfloat &tempZ) {
         m_BBox[tempIndex].min.z = tempZ;
     }
     void SetNoBoundingBoxes(const int &tempSize);
@@ -80,22 +80,22 @@ class AABB {
     //----------------------------------------------------------------------------------
     //  Get Methods
     //----------------------------------------------------------------------------------
-    GLdouble GetMaxX(const int &tempIndex) {
+    GLfloat GetMaxX(const int &tempIndex) {
         return m_BBox[tempIndex].max.x;
     }
-    GLdouble GetMinX(const int &tempIndex) {
+    GLfloat GetMinX(const int &tempIndex) {
         return m_BBox[tempIndex].min.x;
     }
-    GLdouble GetMaxY(const int &tempIndex) {
+    GLfloat GetMaxY(const int &tempIndex) {
         return m_BBox[tempIndex].max.y;
     }
-    GLdouble GetMinY(const int &tempIndex) {
+    GLfloat GetMinY(const int &tempIndex) {
         return m_BBox[tempIndex].min.y;
     }
-    GLdouble GetMaxZ(const int &tempIndex) {
+    GLfloat GetMaxZ(const int &tempIndex) {
         return m_BBox[tempIndex].max.z;
     }
-    GLdouble GetMinZ(const int &tempIndex) {
+    GLfloat GetMinZ(const int &tempIndex) {
         return m_BBox[tempIndex].min.z;
     }
     int GetNoBoundingBoxes() {
