+++ src/PlainNode.h
@@ -32,25 +32,25 @@ class PlainNode {
     //----------------------------------------------------------------------------------
     //  Get Methods
     //----------------------------------------------------------------------------------
-    GLdouble GetType() {
+    GLfloat GetType() {
         return m_type;
     }
-    GLdouble GetXstart() {
+    GLfloat GetXstart() {
         return m_xPlainStart;
     }
-    GLdouble GetXend() {
+    GLfloat GetXend() {
         return m_xPlainEnd;
     }
-    GLdouble GetYstart() {
+    GLfloat GetYstart() {
         return m_yPlainStart;
     }
-    GLdouble GetYend() {
+    GLfloat GetYend() {
         return m_yPlainEnd;
     }
-    GLdouble GetZstart() {
+    GLfloat GetZstart() {
         return m_zPlainStart;
     }
-    GLdouble GetZend() {
+    GLfloat GetZend() {
         return m_zPlainEnd;
     }
 
@@ -62,9 +62,9 @@ class PlainNode {
     //----------------------------------------------------------------------------------
     //  Set Methods
     //----------------------------------------------------------------------------------
-    void SetData(const int tempType, const GLdouble tempXs, const GLdouble tempXe,
-                 const GLdouble tempYs, const GLdouble tempYe,
-                 const GLdouble tempZs, const GLdouble tempZe);
+    void SetData(const int tempType, const GLfloat tempXs, const GLfloat tempXe,
+                 const GLfloat tempYs, const GLfloat tempYe,
+                 const GLfloat tempZs, const GLfloat tempZe);
 
     // Set the address of the link to the next node in the list
     void SetNext(PlainNode *next) {
@@ -79,12 +79,12 @@ class PlainNode {
     // (0: flat plane)
     // (1: incline from z to y)
     // (2: incline from x to y)
-    GLdouble m_type = 0;
+    GLfloat m_type = 0;
 
     // stores start and end co-ordinates of plane on x, y and z axis
-    GLdouble m_xPlainStart = 0, m_xPlainEnd = 0;
-    GLdouble m_yPlainStart = 0, m_yPlainEnd = 0;
-    GLdouble m_zPlainStart = 0, m_zPlainEnd = 0;
+    GLfloat m_xPlainStart = 0, m_xPlainEnd = 0;
+    GLfloat m_yPlainStart = 0, m_yPlainEnd = 0;
+    GLfloat m_zPlainStart = 0, m_zPlainEnd = 0;
 
     //----------------------------------------------------------------------------------
